Dataset added

+++ tumor_img.py
@@ -53,7 +53,7 @@ def teachable_machine_classification2(img, weights_file):
     model = keras.models.load_model(weights_file)
 
     # Create the array of the right shape to feed into the keras model
-    data = np.ndarray(shape=(1, 224, 224, 3), dtype=np.float64)
+    data = np.ndarray(shape=(1, 224, 224, 3), dtype=np.float32)
    
     image = img
     #image sizing
